Split up slices; fix back button label in iOS
