+++ src/components/SharedComponents/Buttons/BackButton.js
@@ -38,7 +38,8 @@ const BackButton = ( {
   if ( navigation?.canGoBack( ) ) {
     return (
       <HeaderBackButton
-        tintColor={tintColor}
+        accessibilityLabel={t( "Navigate-back" )}
+        labelVisible={false}
         onPress={onPress || navigation.goBack}
         // move backbutton to same start as in react-navigation
         style={[
@@ -46,7 +47,7 @@ const BackButton = ( {
           customStyles
         ]}
         testID={testID}
-        accessibilityLabel={t( "Navigate-back" )}
+        tintColor={tintColor}
       />
     );
   }
