+++ src/stores/createExploreSlice.js
@@ -1,0 +1,6 @@
+const createExploreSlice = set => ( {
+  rootExploreParams: {},
+  setRootExploreParams: params => set( ( ) => ( { rootExploreParams: params } ) )
+} );
+
+export default createExploreSlice;
